wallet: add no-sync source
This adds a new chain sync source that simply doesn't sync the wallet.
When passing this there's no need to connect to an external chain
source, which makes everything easier to wire up. Useful in integration
test scenarios!

Note that the wallet is still updated on new block events received via
Bitcoin Core.