main: verify txindex exists on wallet startup
Currently the wallet assumes that the underlying node has `txindex`
enabled. If this is not the case we'll fail later with cryptic error
messages. Better to fail early!

It could very well be that the best choice here is to _remove_ the need
for a TX index. That's a bigger task that can be done at a later point
if needed.