dev: check import sorts with `cargo fmt` (#204)
* dev: check import sorts with `cargo fmt`

We have to use the nightly Rust version for this, as the relevant
options in `rustfmt.toml` are not stable. They have been very much
stable (!) in this situation for years, but there are some outstanding
issues that prevent it from making its way into stable releases.

* dev: forbid unqualified local imports with Clippy

This requires mixing in nightly Clippy. Not the most beatiful thing, but
it works!

* ci: split nightly/stable clippy checks into two jobs