validator: implement fetching CDiskBlockIndex from LevelDB
LevelDB uses a non-trivial file format, so we pull in a dependency for
this. The dependency is in pure Rust, so we avoid having to depend on
system libraries.

The intention is to fetch CDiskBlockIndex structs from LevelDB to
determine where we should start parsing block files. This is in other
words only an operation we would expect to do a few times (maybe only
even once).

To get around limitations on locking behavior (only a single user of a
LevelDB instance is allowed at a time, and Bitcoin Core is that user) we
copy over the entire LevelDB directory before reading it. This is around
150 MB as of implementation. The whole operation (copying, reading,
wiping) takes sub 100ms on a 2023 MacBook Pro, reading data from an
external SSD. Sounds completely acceptable, in other words!